fix: Game: moved various functions to different TUs
* CreateFinalScoreAndEndGameDataFile() moved from game to statistics TU.
* receive_input_player_name() and make_finalscore_display_data() moved
from anonymous namespace from game to statistics TU.
* AskForPlayerNamePrompt() and MessageScoreSavedPrompt() moved from
game-graphics to statistics-graphics TU.
* finalscore_display_data_t{} and EndGameStatisticsPrompt() moved from
game-graphics to scores-graphics TU.
