+++ src/game-graphics.cpp
@@ -25,25 +25,6 @@ std::string AsciiArt2048() {
   return title_card_richtext.str();
 }
 
-std::string MessageScoreSavedPrompt() {
-  constexpr auto score_saved_text = "Score saved!";
-  constexpr auto sp = "  ";
-  std::ostringstream score_saved_richtext;
-  score_saved_richtext << "\n"
-                       << green << bold_on << sp << score_saved_text << bold_off
-                       << def << "\n";
-  return score_saved_richtext.str();
-}
-
-std::string AskForPlayerNamePrompt() {
-  constexpr auto score_prompt_text =
-      "Please enter your name to save this score: ";
-  constexpr auto sp = "  ";
-  std::ostringstream score_prompt_richtext;
-  score_prompt_richtext << bold_on << sp << score_prompt_text << bold_off;
-  return score_prompt_richtext.str();
-}
-
 std::string BoardInputPrompt() {
   const auto board_size_prompt_text = {
       "(NOTE: Scores and statistics will be saved only for the 4x4 gameboard)\n",
@@ -166,49 +147,6 @@ std::string InputCommandListFooterPrompt() {
   for (const auto txt : input_commands_list_footer_text) {
     str_os << sp << txt << "\n";
   }
-  return str_os.str();
-}
-
-std::string EndGameStatisticsPrompt(finalscore_display_data_t finalscore) {
-  std::ostringstream str_os;
-  constexpr auto stats_title_text = "STATISTICS";
-  constexpr auto divider_text = "──────────";
-  constexpr auto sp = "  ";
-  const auto stats_attributes_text = {
-      "Final score:", "Largest Tile:", "Number of moves:", "Time taken:"};
-  enum FinalScoreDisplayDataFields {
-    IDX_FINAL_SCORE_VALUE,
-    IDX_LARGEST_TILE,
-    IDX_MOVE_COUNT,
-    IDX_DURATION,
-    MAX_NUM_OF_FINALSCOREDISPLAYDATA_INDEXES
-  };
-  const auto data_stats =
-      std::array<std::string, MAX_NUM_OF_FINALSCOREDISPLAYDATA_INDEXES>{
-          std::get<IDX_FINAL_SCORE_VALUE>(finalscore),
-          std::get<IDX_LARGEST_TILE>(finalscore),
-          std::get<IDX_MOVE_COUNT>(finalscore),
-          std::get<IDX_DURATION>(finalscore)};
-
-  std::ostringstream stats_richtext;
-  stats_richtext << yellow << sp << stats_title_text << def << "\n";
-  stats_richtext << yellow << sp << divider_text << def << "\n";
-
-  auto counter{0};
-  const auto populate_stats_info = [data_stats, stats_attributes_text, &counter,
-                                    &stats_richtext](const std::string) {
-    stats_richtext << sp << std::left << std::setw(19)
-                   << std::begin(stats_attributes_text)[counter] << bold_on
-                   << std::begin(data_stats)[counter] << bold_off << "\n";
-    counter++;
-  };
-
-  for (const auto s : stats_attributes_text) {
-    populate_stats_info(s);
-  }
-
-  str_os << stats_richtext.str();
-  str_os << "\n\n";
   return str_os.str();
 }
