+++ src/headers/game-graphics.hpp
@@ -13,8 +13,6 @@ enum { COMPETITION_GAME_BOARD_PLAY_SIZE = 4 };
 namespace Game {
 namespace Graphics {
 std::string AsciiArt2048();
-std::string MessageScoreSavedPrompt();
-std::string AskForPlayerNamePrompt();
 std::string BoardInputPrompt();
 std::string YouWinPrompt();
 std::string GameOverPrompt();
@@ -27,9 +25,6 @@ std::string BoardSizeErrorPrompt();
 std::string InputCommandListPrompt();
 std::string EndlessModeCommandListPrompt();
 std::string InputCommandListFooterPrompt();
-using finalscore_display_data_t =
-    std::tuple<std::string, std::string, std::string, std::string>;
-std::string EndGameStatisticsPrompt(finalscore_display_data_t finalscore);
 using scoreboard_display_data_t =
     std::tuple<bool, std::string, std::string, std::string>;
 std::string GameScoreBoardBox(scoreboard_display_data_t scdd);
