+++ src/headers/scores-graphics.hpp
@@ -12,6 +12,10 @@ using scoreboard_display_data_t =
 
 using scoreboard_display_data_list_t = std::vector<scoreboard_display_data_t>;
 std::string ScoreboardOverlay(scoreboard_display_data_list_t sbddl);
+
+using finalscore_display_data_t =
+    std::tuple<std::string, std::string, std::string, std::string>;
+std::string EndGameStatisticsPrompt(finalscore_display_data_t finalscore);
 } // namespace Graphics
 } // namespace Scoreboard
