+++ src/headers/statistics.hpp
@@ -24,6 +24,7 @@ using load_stats_status_t = std::tuple<bool, total_game_stats_t>;
 load_stats_status_t loadFromFileStatistics(std::string filename);
 ull load_game_best_score();
 void saveEndGameStats(Scoreboard::Score finalscore);
+void CreateFinalScoreAndEndGameDataFile(Scoreboard::Score finalscore);
 } // namespace Statistics
 
 std::istream &operator>>(std::istream &is, Statistics::total_game_stats_t &s);
