+++ src/scores-graphics.cpp
@@ -66,6 +66,49 @@ std::string ScoreboardOverlay(scoreboard_display_data_list_t sbddl) {
   return str_os.str();
 }
 
+std::string EndGameStatisticsPrompt(finalscore_display_data_t finalscore) {
+  std::ostringstream str_os;
+  constexpr auto stats_title_text = "STATISTICS";
+  constexpr auto divider_text = "──────────";
+  constexpr auto sp = "  ";
+  const auto stats_attributes_text = {
+      "Final score:", "Largest Tile:", "Number of moves:", "Time taken:"};
+  enum FinalScoreDisplayDataFields {
+    IDX_FINAL_SCORE_VALUE,
+    IDX_LARGEST_TILE,
+    IDX_MOVE_COUNT,
+    IDX_DURATION,
+    MAX_NUM_OF_FINALSCOREDISPLAYDATA_INDEXES
+  };
+  const auto data_stats =
+      std::array<std::string, MAX_NUM_OF_FINALSCOREDISPLAYDATA_INDEXES>{
+          std::get<IDX_FINAL_SCORE_VALUE>(finalscore),
+          std::get<IDX_LARGEST_TILE>(finalscore),
+          std::get<IDX_MOVE_COUNT>(finalscore),
+          std::get<IDX_DURATION>(finalscore)};
+
+  std::ostringstream stats_richtext;
+  stats_richtext << yellow << sp << stats_title_text << def << "\n";
+  stats_richtext << yellow << sp << divider_text << def << "\n";
+
+  auto counter{0};
+  const auto populate_stats_info = [data_stats, stats_attributes_text, &counter,
+                                    &stats_richtext](const std::string) {
+    stats_richtext << sp << std::left << std::setw(19)
+                   << std::begin(stats_attributes_text)[counter] << bold_on
+                   << std::begin(data_stats)[counter] << bold_off << "\n";
+    counter++;
+  };
+
+  for (const auto s : stats_attributes_text) {
+    populate_stats_info(s);
+  }
+
+  str_os << stats_richtext.str();
+  str_os << "\n\n";
+  return str_os.str();
+}
+
 } // namespace Graphics
 
 } // namespace Scoreboard
